+++ test/calfpath/route_test.clj
@@ -194,3 +194,35 @@ Available URI templates:
     (routes-helper (r/make-dispatcher final-partial-routes) partial-400))
   (testing "unrolled partial (path params)"
     (pp-routes-helper (r/make-dispatcher pp-final-partial-routes) partial-400)))
+
+
+(def flat-routes
+  [{:uri "/info/:token/"            :method :get  :name "info"}
+   {:uri "/album/:lid/artist/:rid/" :method :get }
+   {:uri "/user/:id/profile/:type/" :nested [{:method :get     :name "get.user.profile"}
+                                             {:method :patch   :name "update.user.profile"}
+                                             {:method :delete  :name "delete.user.profile"}]}
+   {:uri "/user/:id/permissions/"   :nested [{:method :get     :name "get.user.permissions"}
+                                             {:method :post    :name "create.user.permission"}
+                                             {:method :put     :name "replace.user.permissions"}]}
+   {:uri "/user/:id/auth"           }
+   {:uri "/hello/1234/"             }])
+
+
+(def trie-routes
+  [{:uri "/info/:token/"            :method :get  :name "info"}
+   {:uri "/album/:lid/artist/:rid/" :method :get }
+   {:uri "/user/:id/*" :nested [{:uri "profile/:type/" :nested [{:method :get     :name "get.user.profile"}
+                                                                {:method :patch   :name "update.user.profile"}
+                                                                {:method :delete  :name "delete.user.profile"}]}
+                                {:uri "permissions/"   :nested [{:method :get     :name "get.user.permissions"}
+                                                                {:method :post    :name "create.user.permission"}
+                                                                {:method :put     :name "replace.user.permissions"}]}
+                                {:uri "auth"           }]}
+   {:uri "/hello/1234/"             }])
+
+
+(deftest test-routes->wildcard-trie
+  (is (= trie-routes
+        (-> flat-routes
+          (r/update-routes r/routes->wildcard-trie {:trie-threshold 2})))))